moved tree walker into its own file and generate ROOT nodes as well
