+++ radio/src/storage/yaml/yaml_datastructs.h
@@ -1,0 +1,9 @@
+#ifndef _YAML_DATASTRUCTS_H_
+#define _YAML_DATASTRUCTS_H_
+
+struct YamlNode;
+
+const YamlNode* get_radiodata_nodes();
+const YamlNode* get_modeldata_nodes();
+
+#endif
