+++ radio/src/storage/yaml/yaml_node.h
@@ -26,16 +26,16 @@ struct YamlIdStr
     const char*  str;
 };
 
+// return false if error
+typedef bool (*yaml_writer_func)(void* opaque, const char* str, size_t len);
+
 struct YamlNode
 {
     typedef bool (*is_active_func)(uint8_t* data);
 
     typedef uint32_t (*cust_to_uint_func)(const YamlNode* node, const char* val, uint8_t val_len);
 
-    // return false if error
-    typedef bool (*writer_func)(void* opaque, const char* str, size_t len);
-
-    typedef bool (*uint_to_cust_func)(const YamlNode* node, uint32_t val, writer_func wf, void* opaque);
+    typedef bool (*uint_to_cust_func)(const YamlNode* node, uint32_t val, yaml_writer_func wf, void* opaque);
 
     typedef uint8_t (*select_member_func)(uint8_t* data);
     
@@ -113,104 +113,5 @@ struct YamlNode
             }}                                                          \
     }
 
-
-class YamlTreeWalker
-{
-    struct State {
-        const YamlNode* node;
-        uint32_t    bit_ofs;
-        int8_t      attr_idx;
-        uint16_t    elmts;
-
-        inline uint32_t getOfs() {
-            return bit_ofs + node->size * elmts;
-        }
-    };
-
-    State   stack[NODE_STACK_DEPTH];
-    uint8_t stack_level;
-    uint8_t virt_level;
-    uint8_t anon_union;
-
-    uint8_t* data;
-
-    uint32_t getAttrOfs() { return stack[stack_level].bit_ofs; }
-    uint32_t getLevelOfs() {
-        if (stack_level < NODE_STACK_DEPTH - 1) {
-            return stack[stack_level + 1].getOfs();
-        }
-        return 0;
-    }
-
-    void setNode(const YamlNode* node) { stack[stack_level].node = node; }
-    void setAttrIdx(uint8_t idx) { stack[stack_level].attr_idx = idx; }
-
-    void setAttrOfs(unsigned int ofs) { stack[stack_level].bit_ofs = ofs; }
-
-    void incAttr() { stack[stack_level].attr_idx++; }
-    void incElmts() { stack[stack_level].elmts++; }
-
-    bool empty() { return stack_level == NODE_STACK_DEPTH; }
-    bool full()  { return stack_level == 0; }
-    
-    // return true on success
-    bool push();
-    bool pop();
-    
-    // Rewind to the current node's first attribute
-    // (and reset the bit offset)
-    void rewind();
-
-public:
-    YamlTreeWalker();
-
-    void reset(const YamlNode* node, uint8_t* data);
-
-    int getLevel() {
-        return NODE_STACK_DEPTH - stack_level
-            + virt_level - anon_union;
-    }
-    
-    const YamlNode* getNode() {
-        return stack[stack_level].node;
-    }
-
-    const YamlNode* getAttr() {
-        int8_t idx = stack[stack_level].attr_idx;
-        if (idx >= 0)
-            return &(stack[stack_level].node->u._array.child[idx]);
-
-        return NULL;
-    }
-
-    uint16_t getElmts() {
-        return stack[stack_level].elmts;
-    }
-
-    // Increment the cursor until a match is found or the end of
-    // the current collection (node of type YDT_NONE) is reached.
-    //
-    // return true if a match has been found.
-    bool findNode(const char* tag, uint8_t tag_len);
-
-    // Get the current bit offset
-    unsigned int getBitOffset();
-
-    bool toParent();
-    bool toChild();
-
-    bool toNextElmt();
-    void toNextAttr();
-
-    bool isElmtEmpty(uint8_t* data);
-
-    void setAttrValue(char* buf, uint8_t len);
-
-    bool generate(YamlNode::writer_func wf, void* opaque);
-
-    void dump_stack();
-};
-
-extern const YamlParserCalls YamlTreeWalkerCalls;
 
 #endif
